+++ server/src/main/java/com/yeohaeng_ttukttak/server/application/member/controller/MemberController.java
@@ -6,7 +6,7 @@ import com.yeohaeng_ttukttak.server.application.member.controller.dto.UpdateMemb
 import com.yeohaeng_ttukttak.server.application.member.service.FindMemberService;
 import com.yeohaeng_ttukttak.server.application.member.service.MemberTravelService;
 import com.yeohaeng_ttukttak.server.application.member.service.UpdateMemberService;
-import com.yeohaeng_ttukttak.server.common.aop.annotation.Authentication;
+import com.yeohaeng_ttukttak.server.common.authentication.Authentication;
 import com.yeohaeng_ttukttak.server.common.dto.ServerResponse;
 import com.yeohaeng_ttukttak.server.domain.auth.dto.AuthenticationContext;
 import com.yeohaeng_ttukttak.server.domain.member.dto.MemberDto;
