+++ server/src/main/java/com/yeohaeng_ttukttak/server/application/travel_city/TravelCityController.java
@@ -3,9 +3,8 @@ package com.yeohaeng_ttukttak.server.application.travel_city;
 import com.yeohaeng_ttukttak.server.application.travel.service.TravelService;
 import com.yeohaeng_ttukttak.server.application.travel_city.dto.AddTravelCityResponse;
 import com.yeohaeng_ttukttak.server.application.travel_city.dto.CityListResponse;
-import com.yeohaeng_ttukttak.server.common.aop.annotation.Authentication;
+import com.yeohaeng_ttukttak.server.common.authentication.Authentication;
 import com.yeohaeng_ttukttak.server.common.dto.ServerResponse;
-import com.yeohaeng_ttukttak.server.domain.auth.dto.AuthenticationContext;
 import com.yeohaeng_ttukttak.server.domain.geography.dto.GeographyDto;
 import com.yeohaeng_ttukttak.server.domain.travel.dto.TravelDto;
 import lombok.RequiredArgsConstructor;
