+++ server/src/main/java/com/yeohaeng_ttukttak/server/application/travel_plan_comment/TravelPlanCommentController.java
@@ -2,7 +2,7 @@ package com.yeohaeng_ttukttak.server.application.travel_plan_comment;
 
 import com.yeohaeng_ttukttak.server.application.travel_plan_comment.dto.CommentListResponse;
 import com.yeohaeng_ttukttak.server.application.travel_plan_comment.dto.CommentContentRequest;
-import com.yeohaeng_ttukttak.server.common.aop.annotation.Authentication;
+import com.yeohaeng_ttukttak.server.common.authentication.Authentication;
 import com.yeohaeng_ttukttak.server.common.dto.ServerResponse;
 import com.yeohaeng_ttukttak.server.domain.auth.dto.AuthenticationContext;
 import com.yeohaeng_ttukttak.server.domain.comment.CommentDto;
