+++ server/src/main/java/com/yeohaeng_ttukttak/server/common/authentication/Authentication.java
@@ -1,4 +1,4 @@
-package com.yeohaeng_ttukttak.server.common.aop.annotation;
+package com.yeohaeng_ttukttak.server.common.authentication;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
