+++ server/src/main/java/com/yeohaeng_ttukttak/server/common/authorization/AuthorizationConfig.java
@@ -1,0 +1,14 @@
+package com.yeohaeng_ttukttak.server.common.authorization;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class AuthorizationConfig {
+
+    @Bean
+    public RoleBasedPermissionManager permissionManagerFactory() {
+        return new RoleBasedPermissionManager();
+    }
+
+}
