+++ server/src/main/java/com/yeohaeng_ttukttak/server/common/authorization/interfaces/Authorizable.java
@@ -1,0 +1,3 @@
+package com.yeohaeng_ttukttak.server.common.authorization.interfaces;
+
+public interface Authorizable {}
