--- server/src/main/java/com/yeohaeng_ttukttak/server/domain/shared/interfaces/Authorizable.java
@@ -1,3 +1,0 @@
-package com.yeohaeng_ttukttak.server.domain.shared.interfaces;
-
-public interface Authorizable {}
